requirements.txt
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
anyio==4.9.0
APScheduler==3.11.0
attrs==25.3.0
cachetools==4.2.2
ccxt==4.4.92
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
frozenlist==1.7.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
gspread==6.2.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
multidict==6.6.3
numpy==2.3.1
oauth2client==4.1.3
oauthlib==3.3.1
pandas==2.3.1
propcache==0.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycares==4.9.0
pycparser==2.22
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-telegram-bot==22.2
pytz==2025.2
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
ta==0.11.0
tornado==6.5.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
yarl==1.20.1
psycopg2-binary==2.9.9